Separate nodes relationship content creator
