+++ data/events/parallel-2-it-eu-institutions.md
@@ -6,10 +6,9 @@ ends: '15:15'
 venue: 'Parallel session, room to be announced'
 color: grey
 register: false
-readMore: false
+readMore: true
 order: '3'
 speakers: []
 teams:
-  - teams: INNO
+  - team: INNO
 ---
-
